+++ packages/dynamic-bonding-curve/tests/buildCurve.test.ts
@@ -3,6 +3,7 @@ import { buildCurve } from '../src/helpers'
 import BN from 'bn.js'
 import {
     ActivationType,
+    BuildCurveBaseParam,
     CollectFeeMode,
     FeeSchedulerMode,
     MigrationFeeOption,
@@ -13,7 +14,7 @@ import {
 import { convertBNToDecimal } from './utils/common'
 
 describe('buildCurve tests', () => {
-    const baseParams = {
+    const baseParams: BuildCurveBaseParam = {
         totalTokenSupply: 1000000000,
         migrationOption: MigrationOption.MET_DAMM_V2,
         tokenBaseDecimal: TokenDecimal.SIX,
@@ -43,6 +44,11 @@ describe('buildCurve tests', () => {
         creatorLockedLpPercentage: 0,
         creatorTradingFeePercentage: 0,
         leftover: 10000,
+        tokenUpdateAuthority: 0,
+        migrationFee: {
+            feePercentage: 0,
+            creatorFeePercentage: 0,
+        },
     }
 
     test('build curve with percentage and threshold parameters', () => {