+++ packages/dynamic-bonding-curve/tests/buildCurveWithMarketCap.test.ts
@@ -7,6 +7,7 @@ import {
 import BN from 'bn.js'
 import {
     ActivationType,
+    BuildCurveBaseParam,
     CollectFeeMode,
     FeeSchedulerMode,
     MigrationFeeOption,
@@ -17,7 +18,7 @@ import {
 import { convertBNToDecimal } from './utils/common'
 
 describe('buildCurveWithMarketCap tests', () => {
-    const baseParams = {
+    const baseParams: BuildCurveBaseParam = {
         totalTokenSupply: 1000000000,
         migrationOption: MigrationOption.MET_DAMM_V2,
         tokenBaseDecimal: TokenDecimal.SIX,
@@ -47,6 +48,11 @@ describe('buildCurveWithMarketCap tests', () => {
         creatorLockedLpPercentage: 0,
         creatorTradingFeePercentage: 0,
         leftover: 10000,
+        tokenUpdateAuthority: 0,
+        migrationFee: {
+            feePercentage: 10,
+            creatorFeePercentage: 50,
+        },
     }
 
     test('build curve by market cap 1', () => {
@@ -120,8 +126,7 @@ describe('buildCurveWithMarketCap tests', () => {
             lockedVestingParams.lockedVestingParam.totalVestingDuration,
             lockedVestingParams.lockedVestingParam
                 .cliffDurationFromMigrationTime,
-            lockedVestingParams.tokenBaseDecimal,
-            lockedVestingParams.activationType
+            lockedVestingParams.tokenBaseDecimal
         )
 
         console.log('lockedVesting', convertBNToDecimal(lockedVesting))